+++ frontend/plugins/frontline_ui/src/modules/inbox/constants/conversationsConstants.tsx
@@ -1,0 +1,1 @@
+export const CONVERSATIONS_LIMIT = 24;
