+++ frontend/plugins/frontline_ui/src/modules/inbox/conversations/components/ConversationItem.tsx
@@ -1,0 +1,195 @@
+import {
+  BlockEditorReadOnly,
+  Button,
+  Checkbox,
+  cn,
+  RelativeDateDisplay,
+  useMultiQueryState,
+  useQueryState,
+} from 'erxes-ui';
+import { useConversationContext } from '../hooks/useConversationContext';
+import { useIntegrationDetail } from '@/integrations/hooks/useIntegrations';
+import { BrandsInline, currentUserState, CustomerInline } from 'ui-modules';
+import { useAtomValue, useSetAtom } from 'jotai';
+import { activeConversationState } from '../states/activeConversationState';
+import { ConversationIntegrationBadge } from '@/integrations/components/IntegrationBadge';
+import { useEffect, useState } from 'react';
+import {
+  selectConversationsState,
+  setSelectConversationsState,
+} from '../states/selectConversationsState';
+
+export const ConversationItem = () => {
+  const [conversationId] = useQueryState<string>('conversationId');
+  const [detailView] = useQueryState<boolean>('detailView');
+
+  const { createdAt, updatedAt, customer, integrationId } =
+    useConversationContext();
+
+  const { integration } = useIntegrationDetail({
+    variables: {
+      _id: integrationId,
+    },
+  });
+  const { brandId } = integration || {};
+
+  if (conversationId || detailView) {
+    return (
+      <ConversationContainer className="p-4 pl-6 h-auto overflow-hidden flex-col items-start">
+        <CustomerInline.Provider customer={customer}>
+          <div className="flex w-full gap-3 leading-tight">
+            <ConversationSelector />
+            <div className="flex-1 space-y-1 truncate">
+              <CustomerInline.Title className="truncate" />
+              <div className="font-normal text-accent-foreground text-xs">
+                <BrandsInline
+                  brandIds={[brandId || '']}
+                  placeholder={brandId ? 'brand not found' : 'no brand'}
+                />
+              </div>
+            </div>
+            <div className="ml-auto text-accent-foreground font-medium">
+              <RelativeDateDisplay.Value value={updatedAt || createdAt} />
+            </div>
+          </div>
+          <ConversationItemContent />
+        </CustomerInline.Provider>
+      </ConversationContainer>
+    );
+  }
+
+  return (
+    <ConversationContainer>
+      <CustomerInline.Provider customer={customer}>
+        <ConversationSelector />
+        <CustomerInline.Title className="w-56 truncate flex-none text-foreground" />
+        <ConversationItemContent />
+        <div className="ml-auto font-medium text-accent-foreground w-32 truncate flex-none">
+          to <BrandsInline brandIds={[brandId || '']} />
+        </div>
+        <div className="w-32 text-right flex-none">
+          <RelativeDateDisplay value={updatedAt || createdAt}>
+            <RelativeDateDisplay.Value value={updatedAt || createdAt} />
+          </RelativeDateDisplay>
+        </div>
+      </CustomerInline.Provider>
+    </ConversationContainer>
+  );
+};
+
+export const ConversationItemContent = () => {
+  const { content } = useConversationContext();
+  return (
+    <div className="truncate w-full h-4 [&_*]:text-sm [&_*]:leading-tight [&_*]:font-medium">
+      <BlockEditorReadOnly content={content} />
+    </div>
+  );
+};
+
+const ConversationContainer = ({
+  children,
+  className,
+}: {
+  children: React.ReactNode;
+  className?: string;
+}) => {
+  const [{ conversationId }, setValues] = useMultiQueryState<{
+    conversationId: string;
+    detailView: boolean;
+  }>(['conversationId', 'detailView']);
+  const setActiveConversation = useSetAtom(activeConversationState);
+  const conversation = useConversationContext();
+  const { _id, readUserIds } = conversation || {};
+  const currentUser = useAtomValue(currentUserState);
+  const isRead = readUserIds?.includes(currentUser?._id || '');
+
+  return (
+    <Button
+      key={_id}
+      variant={isRead ? 'secondary' : 'ghost'}
+      size="lg"
+      className={cn(
+        'flex rounded-none h-10 justify-start px-4 gap-3 hover:bg-primary/5 hover:text-foreground w-full',
+        className,
+        isRead && 'font-medium text-muted-foreground',
+        conversationId === _id &&
+          'bg-primary/10 text-foreground hover:bg-primary/10',
+      )}
+      asChild
+      onClick={() => {
+        setActiveConversation(conversation);
+        setValues({
+          conversationId: _id,
+          detailView: true,
+        });
+      }}
+    >
+      <div>{children}</div>
+    </Button>
+  );
+};
+
+const ConversationSelector = () => {
+  const [conversationId] = useQueryState<string>('conversationId');
+  const { integrationId } = useConversationContext();
+
+  return (
+    <div
+      className={cn(
+        'group grid place-items-center relative size-6',
+        conversationId && 'size-8',
+      )}
+    >
+      <div className="absolute size-full bg-primary/10 rounded-full" />
+      <ConversationCheckbox />
+      <div className="transition-opacity duration-200 relative opacity-100 group-hover:opacity-0 peer-data-[state=checked]:opacity-0">
+        <CustomerInline.Avatar
+          size={conversationId ? 'xl' : 'lg'}
+          className=""
+        />
+        <ConversationIntegrationBadge integrationId={integrationId} />
+      </div>
+    </div>
+  );
+};
+
+const ConversationCheckbox = () => {
+  const { _id } = useConversationContext();
+  const [isChecked, setIsChecked] = useState(false);
+  const setSelectConversations = useSetAtom(setSelectConversationsState);
+
+  return (
+    <>
+      <Checkbox
+        checked={isChecked}
+        className="absolute transition-opacity duration-200 opacity-0 group-hover:opacity-100 data-[state=checked]:opacity-100 z-10"
+        onClick={(e) => e.stopPropagation()}
+        onCheckedChange={() => setSelectConversations(_id)}
+      />
+      <ConversationCheckedEffect
+        isChecked={isChecked}
+        setIsChecked={setIsChecked}
+      />
+    </>
+  );
+};
+
+const ConversationCheckedEffect = ({
+  isChecked,
+  setIsChecked,
+}: {
+  isChecked: boolean;
+  setIsChecked: (isChecked: boolean) => void;
+}) => {
+  const { _id } = useConversationContext();
+  const selectConversations = useAtomValue(selectConversationsState);
+
+  useEffect(() => {
+    if (isChecked !== selectConversations.includes(_id)) {
+      setIsChecked(selectConversations.includes(_id));
+    }
+    // eslint-disable-next-line react-hooks/exhaustive-deps
+  }, [selectConversations]);
+
+  return null;
+};
