+++ frontend/plugins/frontline_ui/src/modules/inbox/conversations/context/ConversationContext.tsx
@@ -1,8 +1,10 @@
 import { createContext } from 'react';
 import { IConversation } from '@/inbox/types/Conversation';
+import { IIntegration } from '@/integrations/types/Integration';
 
 export const ConversationContext = createContext<
   IConversation & {
     loading?: boolean;
+    integration?: IIntegration;
   }
->({} as IConversation & { loading?: boolean });
+>({} as IConversation & { loading?: boolean; integration?: IIntegration });
