+++ frontend/plugins/frontline_ui/src/modules/inbox/conversations/conversation-detail/components/ConversationDetail.tsx
@@ -1,5 +1,5 @@
 import { useAtomValue } from 'jotai';
-import { Separator, useQueryState } from 'erxes-ui';
+import { Separator, Skeleton, useQueryState } from 'erxes-ui';
 
 import { ConversationContext } from '@/inbox/conversations/context/ConversationContext';
 import { ConversationHeader } from './ConversationHeader';
@@ -40,15 +40,17 @@ export const ConversationDetail = () => {
 
   if (loading && !currentConversation && !integrationLoading) {
     return (
-      <div className="relative h-full">
-        <InboxMessagesSkeleton />
+      <div className="flex flex-col">
+        <div className="h-12 border-b flex-none flex items-center px-6">
+          <Skeleton className="w-32 h-4" />
+          <Skeleton className="w-32 h-4 ml-auto" />
+        </div>
+        <div className="relative h-full">
+          <InboxMessagesSkeleton />
+        </div>
       </div>
     );
   }
-
-  // if (!['messenger', 'lead'].includes(integration?.kind)) {
-  //   return <UnderConstruction />;
-  // }
 
   return (
     <div className="flex h-full overflow-hidden">
