+++ frontend/plugins/frontline_ui/src/modules/inbox/conversations/conversation-detail/components/ConversationIntegrationDetail.tsx
@@ -1,8 +1,6 @@
 import { lazy, Suspense } from 'react';
 import { IntegrationKind } from '@/types/Integration';
 import { useConversationContext } from '../hooks/useConversationContext';
-import { InboxMessagesSkeleton } from '@/inbox/components/InboxMessagesSkeleton';
-
 const IMapConversationDetail = lazy(() =>
   import('@/integrations/imap/components/ImapConversationDetail').then(
     (module) => ({ default: module.ImapConversationDetail }),
@@ -25,7 +23,7 @@ export const ConversationIntegrationDetail = () => {
   const { integration } = useConversationContext();
 
   return (
-    <Suspense fallback={<InboxMessagesSkeleton isFetched={false} />}>
+    <Suspense fallback={<div />}>
       {integration?.kind === IntegrationKind.IMAP && <IMapConversationDetail />}
       {integration?.kind === IntegrationKind.FACEBOOK_POST && (
         <FbPostConversationDetail />
