+++ frontend/plugins/frontline_ui/src/modules/inbox/conversations/hooks/useConversations.tsx
@@ -1,11 +1,13 @@
-import { GET_CONVERSATIONS } from '~/modules/inbox/conversations/graphql/queries/getConversations';
+import { GET_CONVERSATIONS } from '@/inbox/conversations/graphql/queries/getConversations';
 import { QueryHookOptions, useQuery } from '@apollo/client';
 import { IConversation } from '../../types/Conversation';
 import {
   EnumCursorDirection,
   ICursorListResponse,
   mergeCursorData,
+  validateFetchMore,
 } from 'erxes-ui';
+import { CONVERSATIONS_LIMIT } from '@/inbox/constants/conversationsConstants';
 
 export const useConversations = (
   options?: QueryHookOptions<ICursorListResponse<IConversation>>,
@@ -22,11 +24,17 @@ export const useConversations = (
   }: {
     direction: EnumCursorDirection;
   }) => {
-    if (!pageInfo?.hasNextPage) return;
+    if (!validateFetchMore({ direction, pageInfo })) {
+      return;
+    }
+
     fetchMore({
       variables: {
-        cursor: pageInfo?.endCursor,
-        limit: 50,
+        cursor:
+          direction === EnumCursorDirection.FORWARD
+            ? pageInfo?.endCursor
+            : pageInfo?.startCursor,
+        limit: CONVERSATIONS_LIMIT,
         direction,
       },
       updateQuery: (prev, { fetchMoreResult }) => {
